Add files via upload
Description: 

This website aims to solve the problem of merch listing, this website aims to unite campus merch sellers, including but not limited to, BU Student Council merch and freelance campus sellers (mainly but not limited to students within a specific campus). 

The website aims to lessen the stress when it comes to merch buying for the people selling as it includes: 

Students names
student ID #  
student phone numbers: 
Campus of origin
Course within the campus 
Order History

upon a student placing an order. Of course once the student receives the goods and offers to comment a response of their satisfactory of the goods, only the name and campus origin will be shown to the public (unless chosen otherwise by the student to keep their confidentiality)

The website also has the freedom to choose who it will cater to, likely which campus they want to provide to, example is a store Only within the Polangui campus and its students, mostly for the convinience of the sellers as many cannot cater to different campuses and some can only cater to 1.

For the safety of our fellow students, included to the accounts held by sellers would be: 

Verified Patches specifically for the USC Students 

Certified Sellers for the small business sellers

This is to ensure the inability of mischevious students from creating Fraud accounts 

Accounts that may be suspected of Fraud and Identity theft of a seller or USC account will be held accountable and possible legal actions will be taken account.

USC would be asked for: 
-list of USC members
- ID Verification
- picture and video proof of personell
-verified letter showcasing personell is approved to be an individual of said USC in their campus
- acceptance of being called for approval of account patch 


Campus USC Sellers will be able to: 
- add personells into their seller page as handlers for orders and view of orders 
- add their products and specifications